Blowfish, now i can finally match pwhashes (thanks sixserver, thanks reddwarf)

+++ server/connection.go
@@ -11,7 +11,7 @@ import (
 
 type Connection struct {
 	conn net.Conn
-	seq  uint16
+	seq  uint32
 	data interface{}
 }
